refactor(bots): rename discord and reddit bots
