+++ src/discord_bot/cogs/support_commands.py
@@ -12,8 +12,8 @@ from mistletoe.markdown_renderer import MarkdownRenderer
 
 # local imports
 from src.common.common import avatar, bot_name, colors, data_dir
-from src.discord.views import DocsCommandView
-from src.discord import cogs_common
+from src.discord_bot.views import DocsCommandView
+from src.discord_bot import cogs_common
 
 
 class SupportCommandsCog(discord.Cog):
